refactor: move components into folder
